+++ .codacy.yml
@@ -1,0 +1,2 @@
+exclude_paths:
+  - "e2e/**"
