+++ package.json
@@ -86,7 +86,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "fflate": "^0.8.3",
+    "fflate": "0.8.3",
     "harper.js": "^2.4.0",
     "hunspell-asm": "^4.0.2",
     "katex": "^0.18.1",
