+++ src/components/settings/ai-section-redaction.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { REDACTED_SECRET } from "@/lib/tauri";
+import { REDACTED_MARKER } from "@/lib/tauri";
 import { editableKeys, withKey, withoutKey, type KeyMap } from "./ai-keys";
 
 function isConfigured(saved: string, isCustom: boolean): boolean {
@@ -12,17 +12,17 @@ function isDirty(input: string, saved: string): boolean {
 
 describe("redacted credentials in Settings", () => {
   it("leaves the key field empty rather than showing the marker", () => {
-    const stored = { openai: REDACTED_SECRET, zai: REDACTED_SECRET };
+    const stored = { openai: REDACTED_MARKER, zai: REDACTED_MARKER };
     expect(editableKeys(stored)).toEqual({ openai: "", zai: "" });
   });
 
   it("keeps the Ollama host visible because it is not redacted", () => {
-    const stored = { ollama: "http://localhost:11434", openrouter: REDACTED_SECRET };
+    const stored = { ollama: "http://localhost:11434", openrouter: REDACTED_MARKER };
     expect(editableKeys(stored).ollama).toBe("http://localhost:11434");
   });
 
   it("still reports a provider with a stored key as connected", () => {
-    expect(isConfigured(REDACTED_SECRET, false)).toBe(true);
+    expect(isConfigured(REDACTED_MARKER, false)).toBe(true);
     expect(isConfigured("", false)).toBe(false);
   });
 
@@ -31,26 +31,26 @@ describe("redacted credentials in Settings", () => {
   });
 
   it("offers no save until the user types a new key", () => {
-    expect(isDirty("", REDACTED_SECRET)).toBe(false);
-    expect(isDirty("sk-new", REDACTED_SECRET)).toBe(true);
+    expect(isDirty("", REDACTED_MARKER)).toBe(false);
+    expect(isDirty("sk-new", REDACTED_MARKER)).toBe(true);
   });
 
   it("cannot save the marker itself as a credential", () => {
-    expect(editableKeys({ openai: REDACTED_SECRET }).openai).toBe("");
-    expect(isDirty(editableKeys({ openai: REDACTED_SECRET }).openai, REDACTED_SECRET)).toBe(false);
+    expect(editableKeys({ openai: REDACTED_MARKER }).openai).toBe("");
+    expect(isDirty(editableKeys({ openai: REDACTED_MARKER }).openai, REDACTED_MARKER)).toBe(false);
   });
 });
 
 describe("saving one provider does not disturb the others", () => {
   const stored: KeyMap = {
-    openai: REDACTED_SECRET,
-    groq: REDACTED_SECRET,
+    openai: REDACTED_MARKER,
+    groq: REDACTED_MARKER,
     ollama: "http://localhost:11434",
   };
 
   it("keeps every other credential as a marker when one key is saved", () => {
     expect(withKey(stored, "groq", "gsk-new")).toEqual({
-      openai: REDACTED_SECRET,
+      openai: REDACTED_MARKER,
       groq: "gsk-new",
       ollama: "http://localhost:11434",
     });
@@ -63,12 +63,12 @@ describe("saving one provider does not disturb the others", () => {
   });
 
   it("adding a first key leaves an unrelated stored one alone", () => {
-    expect(withKey(stored, "mistral", "sk-m").openai).toBe(REDACTED_SECRET);
+    expect(withKey(stored, "mistral", "sk-m").openai).toBe(REDACTED_MARKER);
   });
 
   it("removes only the provider being deleted", () => {
     expect(withoutKey(stored, "openai")).toEqual({
-      groq: REDACTED_SECRET,
+      groq: REDACTED_MARKER,
       ollama: "http://localhost:11434",
     });
   });
@@ -77,6 +77,6 @@ describe("saving one provider does not disturb the others", () => {
     const editable = editableKeys(stored);
     const wrong: KeyMap = { ...editable, groq: "gsk-new" };
     expect(wrong.openai).toBe("");
-    expect(withKey(stored, "groq", "gsk-new").openai).toBe(REDACTED_SECRET);
+    expect(withKey(stored, "groq", "gsk-new").openai).toBe(REDACTED_MARKER);
   });
 });
