+++ src/lib/tauri.ts
@@ -657,7 +657,7 @@ export const mcpRegenerateToken = () => invoke<void>("mcp_regenerate_token");
 export const mcpRegisterTools = (
   tools: { name: string; description: string; inputSchema: unknown }[],
 ) => invoke<void>("mcp_register_tools", { tools });
-export const REDACTED_SECRET = "__stored__";
+export const REDACTED_MARKER = "__stored__";
 export const redactedSecretMarker = () => invoke<string>("redacted_secret_marker");
 
 export interface ProviderModel {
